requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.2.10
attrs==20.1.0
autoflake==1.3.1
Babel==2.8.0
bandit==1.6.2
black==19.10b0
bleach==3.1.5
bumpversion==0.5.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.0.3
coveralls==1.10.0
distlib==0.3.1
Django==3.1
django-environ==0.4.5
djangorestframework==3.11.1
doc8==0.8.0
docopt==0.6.2
docutils==0.16
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
flake8-2020==1.5.0
flake8-bandit==2.1.2
flake8-broken-line==0.1.1
flake8-bugbear==20.1.4
flake8-builtins==1.4.2
flake8-comprehensions==3.2.2
flake8-debugger==3.2.1
flake8-fixme==1.1.1
flake8-isort==2.8.0
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-variables-names==0.0.3
gitdb==4.0.5
GitPython==3.1.7
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
isort==4.3.21
Jinja2==2.11.2
keyring==21.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pathspec==0.8.0
pbr==5.4.5
pep8-naming==0.9.1
pkginfo==1.5.0.1
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.3.5
pytest-cov==2.8.1
pytest-runner==5.2
python-dev-tools==2020.2.5
pytz==2020.1
pyupgrade==1.26.2
PyYAML==5.3.1
readme-renderer==26.0
regex==2020.7.14
requests==2.24.0
requests-toolbelt==0.9.1
restructuredtext-lint==1.3.1
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
stevedore==3.2.0
testfixtures==6.14.1
tokenize-rt==4.0.0
toml==0.10.1
tox==3.14.3
tqdm==4.48.2
twine==3.1.1
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==20.0.31
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.1.0
